Split source database domain modules
